+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 2023-May-1
 
 * PHP 8.{0,1,2}: Updated support for New Relic to include all recent php versions.
-* PHP 8.{0,1,2}: Added [Relay](https://relay.so) support. _([@douwezijlstra-frl](https://github.com/douwezijlstra-frl))_
+* PHP 8.{1,2}: Added [Relay](https://relay.so) support. _([@douwezijlstra-frl](https://github.com/douwezijlstra-frl))_
 
 ***